# docs/swagger.yaml
basePath: /api
definitions:
  dtos.GetProductsResponse:
    properties:
      category:
        type: string
      created_date:
        type: string
      id:
        type: string
      image:
        $ref: '#/definitions/dtos.ProductsImageDto'
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
      vendor_id:
        type: string
    type: object
...
  dtos.IdsToDelete:
    properties:
      ids:
        items:
          type: string
        type: array
    type: object
  dtos.OneProductResponse:
    properties:
      category:
        type: string
      created_date:
        type: string
      description:
        type: string
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/dtos.ProductsImageDto'
        type: array
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
      tags:
        items:
          $ref: '#/definitions/dtos.TagDto'
        type: array
      vendor_id:
        type: string
    type: object
...
  dtos.ProductPatchRequest:
    properties:
      category:
        type: string
      description:
        type: string
      images:
        items:
          type: string
        type: array
      name:
        type: string
      price:
        type: number
      tags:
        items:
          type: string
        type: array
    type: object
  dtos.ProductRequest:
    properties:
      category:
        type: string
      description:
        type: string
      images:
        items:
          type: string
        type: array
      name:
        type: string
      price:
        type: number
      tags:
        items:
          type: string
        type: array
    type: object
  dtos.ProductsImageDto:
    properties:
      id:
        type: string
      image_url:
        type: string
      product_id:
        type: string
    type: object
...
  dtos.TagDto:
    properties:
      id:
        type: string
      tag_name:
        type: string
    type: object
host: marketplace-vendor-products-service.onrender.com
info:
  contact: {}
  description: API for managing products and stocks for vendors.
  title: Product Service API
  version: "1.0"
paths:
  /products:
    delete:
      consumes:
      - application/json
      description: Deletes multiple products for the given vendor by IDs.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: IDs to delete
        in: body
        name: ids
        required: true
        schema:
          $ref: '#/definitions/dtos.IdsToDelete'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid vendorId/IDs of products to delete
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete many products
      tags:
      - products
    get:
      consumes:
      - application/json
      description: Returns a paginated list of products for the given vendor, with
        filtering, searching and sorting options.
      parameters:
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 15
        description: Page size
        in: query
        minimum: 1
        name: size
        type: integer
      - default: -1
        description: Offset
        in: query
        name: offset
        type: integer
      - default: -1
        description: Limit
        in: query
        name: limit
        type: integer
      - description: Category filter
        in: query
        name: category
        type: string
      - default: 0
        description: Minimum price
        in: query
        minimum: 0
        name: minPrice
        type: number
      - default: 0
        description: Maximum price
        in: query
        minimum: 0
        name: maxPrice
        type: number
      - description: Search term
        in: query
        name: search
        type: string
      - default: name
        description: Sort by field
        in: query
        name: sortBy
        type: string
      - default: asc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: sortOrder
        type: string
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.GetProductsResponse'
            type: array
        "400":
          description: Invalid vendorId/page/page size/offset/limit/price
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Creates a new product for the given vendor.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product creation payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.ProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.OneProductResponse'
        "400":
          description: Invalid vendorId/product data
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new product
      tags:
      - products
  /products/id/{productId}:
    delete:
      consumes:
      - application/json
      description: Deletes a single product by its ID for the given vendor.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product ID (UUID)
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid vendorId/productId
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete product by ID
      tags:
      - products
    get:
      consumes:
      - application/json
      description: Returns a single product by its ID for the given vendor.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product ID (UUID)
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.OneProductResponse'
        "400":
          description: Invalid vendorId/productId
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get product by ID
      tags:
      - products
    patch:
      consumes:
      - application/json
      description: Partially updates a product by its ID for the given vendor.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product ID (UUID)
        in: path
        name: productId
        required: true
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.ProductPatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.OneProductResponse'
        "400":
          description: Invalid vendorId/productId/product data
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update product fields by ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Updates an existing product by its ID for the given vendor.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product ID (UUID)
        in: path
        name: productId
        required: true
        type: string
      - description: Full product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.ProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid vendorId/productId/product data
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update product by ID
      tags:
      - products
  /products/slug/{slug}:
    delete:
      consumes:
      - application/json
      description: Deletes a single product by its ID for the given vendor.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product slug (unique identifier in URL)
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid vendorId/slug
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete product by ID
      tags:
      - products
    get:
      consumes:
      - application/json
      description: Returns a single product by its slug for the given vendor.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product slug (unique identifier in URL)
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.OneProductResponse'
        "400":
          description: Invalid vendorId/slug
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get product by slug
      tags:
      - products
    patch:
      consumes:
      - application/json
      description: Partially updates a product by its ID for the given vendor.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product slug (unique identifier in URL)
        in: path
        name: slug
        required: true
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.ProductPatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.OneProductResponse'
        "400":
          description: Invalid vendorId/slug/product data
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update product fields by ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Updates an existing product by its slug for the given vendor.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product slug (unique identifier in URL)
        in: path
        name: slug
        required: true
        type: string
      - description: Full product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.ProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid vendorId/slug/product data
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update product by slug
      tags:
      - products
...
schemes:
- https
swagger: "2.0"